Revert "Out-of-process custom sidebars: isolated interpreter + remote rendering (#5294)"
This reverts commit 56db7a92d633452771b6701696fb9681903a104e.
