--- Packages/CmuxSidebarInterpreterService/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "54fcd85444a80bfa0b5c67ea264b852e871b3bdd84b9a08fc7e1ab77756837a3",
-  "pins" : [
-    {
-      "identity" : "swift-syntax",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/swiftlang/swift-syntax.git",
-      "state" : {
-        "revision" : "0687f71944021d616d34d922343dcef086855920",
-        "version" : "600.0.1"
-      }
-    }
-  ],
-  "version" : 3
-}
