--- Packages/CmuxSidebarInterpreterService/Sources/CmuxSidebarInterpreterClient/InterpreterClient+SidebarInterpreting.swift
@@ -1,6 +1,0 @@
-import CmuxSwiftRender
-
-/// ``InterpreterClient`` is the out-of-process, crash-isolating
-/// ``SidebarInterpreting``: each `render` runs in a supervised worker, so an
-/// interpreter fault returns `nil` instead of crashing the host.
-extension InterpreterClient: SidebarInterpreting {}
