+++ Packages/CmuxSwiftRender/Sources/CmuxSwiftRender/ButtonAction.swift
@@ -1,6 +1,6 @@
 /// The captured behavior of a `Button`, evaluated when the button is tapped
 /// by a host runtime.
-public struct ButtonAction: Codable, Sendable, Equatable {
+public struct ButtonAction: Sendable, Equatable {
     public let commands: [ActionCommand]
 
     public init(commands: [ActionCommand]) {
