+++ Packages/CmuxSwiftRenderUI/Package.swift
@@ -4,7 +4,6 @@ import PackageDescription
 
 let package = Package(
     name: "CmuxSwiftRenderUI",
-    defaultLocalization: "en",
     platforms: [
         .macOS(.v14),
     ],
@@ -26,9 +25,6 @@ let package = Package(
                 .product(name: "CmuxSwiftRender", package: "CmuxSwiftRender"),
                 .product(name: "CmuxSettings", package: "CmuxSettings"),
                 .product(name: "CmuxFileWatch", package: "CmuxFileWatch"),
-            ],
-            resources: [
-                .process("Resources"),
             ]
         ),
         .testTarget(
