--- Packages/CmuxSwiftRenderUI/Sources/CmuxSwiftRenderUI/Rendering/SidebarTapTargetsKey.swift
@@ -1,12 +1,0 @@
-import SwiftUI
-
-/// Accumulates every ``SidebarTapTarget`` in the rendered sidebar.
-public struct SidebarTapTargetsKey: PreferenceKey {
-    /// No targets until the render tree reports some.
-    public static let defaultValue: [SidebarTapTarget] = []
-
-    /// Concatenates targets from sibling subtrees.
-    public static func reduce(value: inout [SidebarTapTarget], nextValue: () -> [SidebarTapTarget]) {
-        value.append(contentsOf: nextValue())
-    }
-}
