--- Packages/CmuxSwiftRenderUI/Sources/CmuxSwiftRenderUI/Resources/Localizable.xcstrings
@@ -1,57 +1,0 @@
-{
-  "sourceLanguage": "en",
-  "strings": {
-    "sidebar.custom.error": {
-      "extractionState": "manual",
-      "localizations": {
-        "en": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "Sidebar error"
-          }
-        },
-        "ja": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "サイドバーエラー"
-          }
-        }
-      }
-    },
-    "sidebar.custom.missing": {
-      "extractionState": "manual",
-      "localizations": {
-        "en": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "Sidebar file is empty or missing."
-          }
-        },
-        "ja": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "サイドバーファイルが空か見つかりません。"
-          }
-        }
-      }
-    },
-    "sidebar.custom.noView": {
-      "extractionState": "manual",
-      "localizations": {
-        "en": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "No supported SwiftUI view found."
-          }
-        },
-        "ja": {
-          "stringUnit": {
-            "state": "translated",
-            "value": "対応するSwiftUIビューが見つかりません。"
-          }
-        }
-      }
-    }
-  },
-  "version": "1.0"
-}
